Add gameclient 12.6 support (#11)
* feat: Add gameclient 12.6.0 assets

This commit updates the gameclient assets for version 12.6.0. The changes also include:
- Modified code in `replay_parser.py` to skip reading and appending extra data
- Updated packet mappings in `__init__.py` for entities, maps, nested properties, position, version, and player position

These caveats should be adressed in future commits

* feat: Backport compatibility for 12.6.0 packet update

This commit backports compatibility for the 12.6.0 packet update by adding a dedicated packet list and updating existing ones in the `__init__.py` file of the `wows.network.packets` module. It also modifies the `player.py` file to handle the updated packet mappings based on the game client version. Additionally, it updates the constructor of the `PlayerBase` class in the `player.py` file to accept a list of strings representing the version number and converts it to a tuple for internal use.

* chore: Finishing touches for 12.6.0 support

- Updated spaces manifest